Working on adding playlist funcationality

+++ audio/converter.py
@@ -24,27 +24,33 @@ def convert_to_webm(song: Song) -> None:
     song.extension = 'webm'
 
 
-def equalise_loudness(song: Song) -> None:
-    # Load the new audio track
-    audio = AudioSegment.from_file(song.full_filename, format="webm")
+def equalise_loudness(song: Song, swap_now) -> None:
+    # Normalized file path
+    normalized_file = song.full_filename_without_extension + '_normalized.webm'
 
-    # Resample the new track if necessary
-    if audio.frame_rate != REFERENCE_SAMPLE_RATE:
-        audio = audio.set_frame_rate(REFERENCE_SAMPLE_RATE)
+    # Check if the normalized file already exists
+    if not os.exists(normalized_file):
+        # Load the new audio track
+        audio = AudioSegment.from_file(song.full_filename, format="webm")
 
-    # Get the loudness of the new track
-    chunks = make_chunks(audio, 1000)
-    loudness = max(chunk.dBFS for chunk in chunks)
+        # Resample the new track if necessary
+        if audio.frame_rate != REFERENCE_SAMPLE_RATE:
+            audio = audio.set_frame_rate(REFERENCE_SAMPLE_RATE)
 
-    # Normalize the loudness of the new track if necessary
-    if loudness != REFERENCE_LOUDNESS:
-        audio = audio.apply_gain(REFERENCE_LOUDNESS - loudness)
+        # Get the loudness of the new track
+        chunks = make_chunks(audio, 1000)
+        loudness = max(chunk.dBFS for chunk in chunks)
 
-    # Export the normalized audio track as a new file
-    temp_file = song.full_filename_without_extension + '_normalized.webm'
-    audio.export(temp_file, format="webm")
-    os.remove(song.full_filename)
-    shutil.move(temp_file, song.full_filename)
+        # Normalize the loudness of the new track if necessary
+        if loudness != REFERENCE_LOUDNESS:
+            audio = audio.apply_gain(REFERENCE_LOUDNESS - loudness)
+
+        # Export the normalized audio track as a new file
+        audio.export(normalized_file, format="webm")
+
+    if swap_now:
+        os.remove(song.full_filename)
+        shutil.move(normalized_file, song.full_filename)
 
 
 def get_loudness(audio_file):
